+++ changelog.md
@@ -5,6 +5,8 @@
 - `mc!dump`, `mc!players` and `mc!pie` have been merged into one command, `mc!csv`. (#303)
 - Participant limits are enforced (#313)
 - Add `mc!info` to display a pretty embed of all relevant information (#314)
+- Add command `mc!capacity` to read and set the participant limit of preparing tournaments (#315)
+
 ### Bug Fixes
 - Usernames printed in Discord will have markdown elements escaped, so they appear as written. (#304)
 - Fix fatal out-of-memory crashes on tournament start (#306)